Added support for choosing connection encoding.
The clojure.contrib.io versions of reader and writer lack options needed
for specifying which encoding should be used. In clojure.java.io, this
feature is present, but there is a bug in the implementation of
output-stream with sockets. Since the java interop calls that the io
library abstracts away are pretty straight forward in this case, I think
that the java interop calls could be used directly until the bug has been fixed.