create-irc now wraps the returned map in a ref. This was previously done by connect.
This change might break things for those who directly access the result of create-irc before passing it to connect. However, the example code and all known running bots (which of most ones are modified version of the example) don't do this, so hopefully this won't be a problem.

connect now also stores the options it gets at the :connect-options key in the bot map. Previously existing options are merged with the new ones. Thus, a connect call with no options will reuse the options used in the last connect call.